+++ kylin-user/src/main/java/com/store59/kylin/user/data/mapper/UserMapper.java
@@ -6,13 +6,13 @@ import com.store59.kylin.user.data.model.User;
 
 public interface UserMapper {
 
-    int deleteByPrimaryKey(Integer uid);
+    int deleteByPrimaryKey(Long uid);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Integer uid);
+    User selectByPrimaryKey(Long uid);
     
     User getUserForLogin(@Param("name")String name,@Param("password")String password);
